requirements.txt
# Core
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# OCR
easyocr>=1.7.0

# Object Detection
ultralytics>=8.0.0

# Face Analysis
mediapipe>=0.10.0

# Token Company
tokenc>=0.1.0

# Token counting
tiktoken>=0.5.0

# LLM clients
openai>=1.0.0
google-generativeai>=0.5.0

# CLI
typer>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0
